New submission
I realized what went wrong in the last submission. I got a new webDev folder from github and I wrote my code in the original folder github gave me. So when I submitted version 6, it was with the original folder, but I don't know how it happened but I got a new folder and github was expecting version 7 in that folder. Sorry for the inconvinience but this should work now.