A lot to version 3.0
Made a bunch of changes to the website most noticeable ones are, that I added. A music tab a list table and made them all a tab you can go into. Music tab as most of my favorite music with the thumbnail in all of them. and the list has my favorite pokemon.